chore: Split core modules
