+++ dev/main/linkdeps.js
@@ -1,12 +1,13 @@
 const { exec } = require('child_process');
 
 [
-  '@garfish/core',
-  '@garfish/router',
-  '@garfish/browser-snapshot',
-  '@garfish/browser-vm',
-  '@garfish/core',
-  '@garfish/loader',
+  '@garfish/framework',
+  // '@garfish/core',
+  // '@garfish/router',
+  // '@garfish/browser-snapshot',
+  // '@garfish/browser-vm',
+  // '@garfish/core',
+  // '@garfish/loader',
 ].forEach((item) => {
   console.log(`link ${item}`);
   exec(`yarn link ${item}`, (error, stdout, stderr) => {
