+++ dev/main/package.json
@@ -22,15 +22,11 @@
     "style-loader": "^1.0.0",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.10",
+    "@garfish/framework": "0.0.8-alpha.5",
     "webpack-dev-server": "^3.9.0"
   },
   "dependencies": {
     "@babel/runtime": "^7.12.1",
-    "@garfish/browser-snapshot": "0.0.8-alpha.5",
-    "@garfish/browser-vm": "0.0.8-alpha.5",
-    "@garfish/cjs-app": "0.0.8-alpha.5",
-    "@garfish/core": "0.0.8-alpha.5",
-    "@garfish/router": "0.0.8-alpha.5",
     "loader-utils": "^2.0.0",
     "mobx": "^6.0.1",
     "react": "^16.13.1",
