+++ dev/subApp/linkdeps.js
@@ -2,11 +2,11 @@ const { exec } = require('child_process');
 
 [
   '@garfish/core',
-  '@garfish/router',
-  '@garfish/browser-snapshot',
-  '@garfish/browser-vm',
-  '@garfish/core',
-  '@garfish/loader',
+  // '@garfish/router',
+  // '@garfish/browser-snapshot',
+  // '@garfish/browser-vm',
+  // '@garfish/core',
+  // '@garfish/loader',
 ].forEach((item) => {
   console.log(`link ${item}`);
   exec(`yarn link ${item}`, (error, stdout, stderr) => {
