+++ dev/subApp/package.json
@@ -13,7 +13,7 @@
     "vue": "^2.6.11",
     "vue-router": "^3.4.3",
     "vuex": "^3.5.1",
-    "@garfish/core": "0.0.8-alpha.5"
+    "@garfish/framework": "0.0.8-alpha.5"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.4.0",
