+++ packages/runtime/garfish/README.md
@@ -1,0 +1,11 @@
+# `@garfish/garfish`
+
+> TODO: description
+
+## Usage
+
+```
+import @garfish/garfish from '@garfish/garfish';
+
+// TODO: DEMONSTRATE API
+```
