+++ packages/runtime/garfish/__tests__/index.spec.ts
@@ -1,0 +1,5 @@
+import toolApp from '../src/index';
+
+describe('garfish', () => {
+  it('needs tests');
+});
