+++ packages/runtime/garfish/api-extractor.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../../api-extractor.json",
+  "mainEntryPointFilePath": "./dist/packages/runtime/garfish/src/index.d.ts",
+  "dtsRollup": {
+    "publicTrimmedFilePath": "./dist/garfish.d.ts"
+  }
+}
