+++ packages/runtime/utils/src/index.ts
@@ -5,3 +5,4 @@ export * from './mimeType';
 export * from './container';
 export * from './webpack';
 export * from './sentry';
+export * from './garfish';
